code is able to train a model

+++ src/dataset.py
@@ -33,8 +33,8 @@ class make_dataset(Dataset):
         token_type_ids = inputs["token_type_ids"]
 
         return {
-            'ids': torch.tensor(ids, dtype=torch.long),
+            'ids': torch.tensor(ids, dtype=torch.float),
             'mask': torch.tensor(mask, dtype=torch.long),
             'token_type_ids': torch.tensor(token_type_ids, dtype=torch.long),
-            'targets': torch.tensor(self.targets[index], dtype=torch.float)
+            'targets': torch.tensor(self.targets[index], dtype=torch.long)
         }
